rename to tai_seconds; all formats in one file; fix docstring
