+++ sunpy/time/__init__.py
@@ -1,4 +1,3 @@
-from sunpy.time.taicds import *
 from sunpy.time.time import *
+from sunpy.time.timeformats import *
 from sunpy.time.timerange import *
-from sunpy.time.utime import *
