+++ sunpy/time/tests/test_taiseconds.py
@@ -4,20 +4,18 @@ from astropy.time import Time
 from astropy.time.formats import erfa
 
 # This registers the TimeTAICDS format with astropy
-from sunpy.time import TimeTAICDS  # NOQA
+from sunpy.time import TimeTAISeconds  # NOQA
 
 
 def test_time_t0():
     """
     Test that the offset at the epoch is zero. Note that the offset
-    will be zero in both TAI and UTC and that this differs from
-    `anytim2tai` which gives a constant offset of 9 s prior to
-    1972.
+    will be zero in both TAI and UTC
     """
     t = Time('1958-01-01 00:00:00', format='iso', scale='tai')
-    assert t.tai.tai_cds == 0.0
+    assert t.tai.tai_seconds == 0.0
     with pytest.warns(erfa.ErfaWarning, match='dubious year'):
-        assert t.utc.tai_cds == 0.0
+        assert t.utc.tai_seconds == 0.0
 
 
 def test_tai_utc_offset():
@@ -27,7 +25,7 @@ def test_tai_utc_offset():
     """
     t1 = Time('1972-01-01 00:00:00', scale='tai', format='iso')
     t2 = Time('1972-01-01 00:00:00', scale='utc', format='iso')
-    assert t2.tai_cds - t1.tai_cds == 10.0
+    assert t2.tai_seconds - t1.tai_seconds == 10.0
 
 
 @pytest.mark.parametrize('time', [
@@ -37,5 +35,5 @@ def test_tai_utc_offset():
     Time('2018-09-17T19:46:25', format='isot', scale='tai'),
 ])
 def test_roundtrip(time):
-    time_from_cds_tai = Time(time.tai_cds, scale='tai', format='tai_cds')
-    assert time_from_cds_tai == time
+    time_from_tai_seconds = Time(time.tai_seconds, scale='tai', format='tai_seconds')
+    assert time_from_tai_seconds == time
